+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/arm/GoScore.java
@@ -1,18 +1,16 @@
-package org.firstinspires.ftc.teamcode.command.lift;
+package org.firstinspires.ftc.teamcode.command.arm;
 import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
 import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 
-public class LiftDown extends SimpleArmCommand {
-
-    public LiftDown(ArmSubsystem arm) {
+public class GoScore extends SimpleArmCommand {
+    public GoScore(ArmSubsystem arm) {
         super(arm);
 
     }
 
     @Override
     public void initialize() {
-//        arm.setLiftManual();
-        arm.LiftDown();
+        arm.armScore();
     }
 }
