+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/arm/PrepToGrabCone.java
@@ -1,18 +1,16 @@
-package org.firstinspires.ftc.teamcode.command.lift;
+package org.firstinspires.ftc.teamcode.command.arm;
 import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
 import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 
-public class LiftUp extends SimpleArmCommand {
-
-    public LiftUp(ArmSubsystem arm) {
+public class PrepToGrabCone extends SimpleArmCommand {
+    public PrepToGrabCone(ArmSubsystem arm) {
         super(arm);
 
     }
 
     @Override
     public void initialize() {
-//        arm.setLiftManual();
-        arm.LiftUp();
+        arm.grabConePrep();
     }
 }
