+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/lift/MoveLiftTo.java
@@ -1,11 +1,10 @@
 package org.firstinspires.ftc.teamcode.command.lift;
 
 import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
-import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 public class MoveLiftTo extends SimpleArmCommand {
     int targetPos;
-    public MoveLiftTo(ArmSubsystem arm, int targetPos) {
+    public MoveLiftTo(ArmSubsystem_OG arm, int targetPos) {
         super(arm);
         this.targetPos = targetPos;
     }
