+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/lift/RunLiftManualy.java
@@ -1,14 +1,13 @@
 package org.firstinspires.ftc.teamcode.command.lift;
 
 import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
-import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 import java.util.function.DoubleSupplier;
 
 public class RunLiftManualy extends SimpleArmCommand {
 
     DoubleSupplier speed;
-    public RunLiftManualy(ArmSubsystem arm, DoubleSupplier speed) {
+    public RunLiftManualy(ArmSubsystem_OG arm, DoubleSupplier speed) {
         super(arm);
         this.speed = speed;
     }
