+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/lift/setAuto.java
@@ -1,11 +1,10 @@
 package org.firstinspires.ftc.teamcode.command.lift;
 import org.firstinspires.ftc.teamcode.command.SimpleArmCommand;
-import org.firstinspires.ftc.teamcode.subsystem.ArmSubsystem;
 
 
 public class setAuto extends SimpleArmCommand {
 
-    public setAuto(ArmSubsystem arm) {
+    public setAuto(ArmSubsystem_OG arm) {
         super(arm);
 
     }
