+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystem/ArmSubsystem.java
@@ -1,27 +1,71 @@
 package org.firstinspires.ftc.teamcode.subsystem;
 
-
+import android.util.Log;
 import com.arcrobotics.ftclib.command.SubsystemBase;
+import com.arcrobotics.ftclib.controller.wpilibcontroller.ProfiledPIDController;
 import com.arcrobotics.ftclib.hardware.ServoEx;
-import com.qualcomm.robotcore.hardware.DcMotor;
+import com.arcrobotics.ftclib.hardware.motors.MotorEx;
+import com.arcrobotics.ftclib.trajectory.TrapezoidProfile;
 
 public class ArmSubsystem extends SubsystemBase{
 
-    private final ServoEx claw;
+    private final ServoEx claw, arm1, arm2;
 
-    private final DcMotor liftL, liftR;
+    private final MotorEx slideL, slideR;
 
-    //TODO tune values appropriately (otherwise no work)
-    private static int NONE = 0;
-    private static int LOW = 0;
-    private static int MEDIUM = 0;
-    private static int HIGH = 0;
-    private static int GROUND = 0;
 
-    //change this to change lift speed
-    double speed = 0.5;
+    //TODO tune height values
+    public static int NONE = 0;
+    public static int GROUND = 0;
+    public static int LOW = 0;
+    public static int MEDIUM = 0;
+    public static int HIGH = 0;
 
-    public enum Junction {
+
+    //PID stuff for motors
+    //TODO NEED TO TUNE THESE VALUES
+    public static double slide_P = 0.003;
+
+    public static double slide_I = 0.05;
+
+    public static double slide_D = 0.0003;
+
+    public static double maxVel = 1000;
+
+    public static double maxAccel = 1000;
+
+    //https://www.ctrlaltftc.com/the-pid-controller/tuning-methods-of-a-pid-controller
+
+    private final ProfiledPIDController slide_pidL = new ProfiledPIDController(slide_P, slide_I, slide_D,
+            new TrapezoidProfile.Constraints(maxVel, maxAccel));
+
+    private final ProfiledPIDController slide_pidR = new ProfiledPIDController(slide_P, slide_I, slide_D,
+            new TrapezoidProfile.Constraints(maxVel, maxAccel));
+
+    private double out_left;
+
+    private double out_right;
+
+    public static double tolerance = 10;
+
+    public static Junction currentGoal = Junction.NONE;
+
+    //TODO TUNE THESE VALUES
+    private final double arm1_home = 0;
+
+    private final double arm2_home = 0;
+
+    private final double arm1_score = 0;
+
+    private final double arm2_score = 0;
+
+    private final double clawOpen = 1;
+
+    private final double clawClose = 0;
+
+    private final double manualSlideSpeed = 0;
+
+    public enum Junction{
         NONE,
         GROUND,
         LOW,
@@ -29,100 +73,127 @@ public class ArmSubsystem extends SubsystemBase{
         HIGH
     }
 
+    //constructor
 
-    public ArmSubsystem(ServoEx claw, DcMotor liftL, DcMotor liftR) {
+    public ArmSubsystem(ServoEx claw, ServoEx arm1, ServoEx arm2, MotorEx slideL, MotorEx slideR){
         this.claw = claw;
-        this.liftL = liftL;
-        this.liftR = liftR;
+        this.arm1 = arm1;
+        this.arm2 = arm2;
+        this.slideL = slideL;
+        this.slideR = slideR;
 
-        liftL.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        liftR.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        slide_pidL.setTolerance(tolerance);
+        slide_pidR.setTolerance(tolerance);
 
-        setLiftPosition(0);
+        slide_pidL.setGoal(0);
+        slide_pidR.setGoal(0);
 
-//        liftL.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-//        liftR.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+    }
 
-        liftL.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        liftR.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+    //claw control
+
+    public void clawOpen() {claw.setPosition(clawOpen);}
+
+    public void clawClose() {claw.setPosition(clawClose);}
+
+
+    //arm control
+
+    public void armHome() {
+        arm1.setPosition(arm1_home);
+        arm2.setPosition(arm2_home);
+    }
+
+    //TODO tune values after servo placed
+    public void armScore() {
+        arm1.setPosition(arm1_score);
+        arm2.setPosition(arm2_score);
+    }
+
+    //would this even work? would both servos have the same motion thingy???
+    public void armPos(double pos){
+        arm1.setPosition(pos);
+        arm2.setPosition(pos);
     }
 
 
-    public void clawOpen() {claw.setPosition(1);}
 
-    public void clawClose() {claw.setPosition(0);}
 
-    public void setLiftPosition(int targetPosition) {
-        liftL.setTargetPosition(targetPosition);
-        liftR.setTargetPosition(targetPosition);
+    //slide pid stuff
+
+    public void loopPID(){
+        out_left = slide_pidL.calculate(slideL.getCurrentPosition());
+        out_right = slide_pidR.calculate(slideR.getCurrentPosition());
+        slideL.set(out_left);
+        slideR.set(out_right);
     }
 
-    public void LiftUp(){
-//        liftL.setPower(speed);
-//        liftR.setPower(speed);
-
-        liftL.setTargetPosition(liftL.getCurrentPosition() - 20);
-        liftR.setTargetPosition(liftR.getCurrentPosition() - 20);
-
-    }
-    public void LiftDown(){
-
-//        liftL.setPower(-speed);
-//        liftR.setPower(-speed);
-
-        liftL.setTargetPosition(liftL.getCurrentPosition() + 20);
-        liftR.setTargetPosition(liftR.getCurrentPosition() + 20);
-
-    }
-    public void LiftStop(){
-
-        liftL.setPower(0);
-        liftR.setPower(0);
+    public void setSlidesPID (int goal){
+        slide_pidL.setGoal(goal);
+        slide_pidR.setGoal(goal);
     }
 
-    public void setLiftManual(){
-        liftL.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        liftR.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-    }
-
-
-    public void setLiftAuto(){
-        liftR.setTargetPosition(liftR.getCurrentPosition());
-        liftL.setTargetPosition(liftL.getCurrentPosition());
-        liftL.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        liftR.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-    }
-
-    public int[] getLiftsPos () {
-        return new int[] {
-                liftL.getCurrentPosition(),
-                liftR.getCurrentPosition()
-        };
-    }
-
-
-    public void setJunction(ArmSubsystem arm, Junction junction) {
-        switch(junction) {
+    public void moveToJunction(Junction junction){
+        currentGoal = junction;
+        switch(junction){
             case NONE:
-                arm.setLiftPosition(0);
-                // arm.setLiftPosition(getLiftsPos()[0]);
+                setSlidesPID(NONE);
                 break;
             case GROUND:
-                arm.setLiftPosition(GROUND);
+                setSlidesPID(GROUND);
                 break;
             case LOW:
-                arm.setLiftPosition(LOW);
+                setSlidesPID(LOW);
                 break;
             case MEDIUM:
-                arm.setLiftPosition(MEDIUM);
+                setSlidesPID(MEDIUM);
                 break;
             case HIGH:
-                arm.setLiftPosition(HIGH);
+                setSlidesPID(HIGH);
+                break;
         }
+        out_left = slide_pidL.calculate(slideL.getCurrentPosition());
+
+        while(!slide_pidL.atGoal()){
+            out_left = out_left = slide_pidL.calculate(slideL.getCurrentPosition());
+            Log.d("output", "" + out_left);
+            Log.d("error", "" + slide_pidL.getPositionError());
+            Log.d("encoder", "" + slideL.getCurrentPosition());
+
+            out_left = out_right;
+
+            slideL.set(out_left);
+            slideR.set(out_right);
+        }
+
+
     }
 
-    public void manualControlLift(double power) {
-        liftL.setPower(power/2);
-        liftR.setPower(power/2);
+    public void changeSetPoint(double input){
+        slide_pidL.setGoal((int)(slideL.getCurrentPosition()+input*manualSlideSpeed));
+        slide_pidL.setGoal((int)(slideL.getCurrentPosition()+input*manualSlideSpeed));
+        Log.d("setpoint left", "" + slide_pidL.getSetpoint().position);
+        Log.d("setpoint right", "" + slide_pidR.getSetpoint().position);
     }
+
+
+
+    public int getSlideLEncoder(){
+        return slideL.getCurrentPosition();
+    }
+    public int getSlideREncoder(){
+        return slideR.getCurrentPosition();
+    }
+
+    public double getSlideLPower(){
+        return out_left;
+    }
+    public double getSlideRPower(){
+        return out_right;
+    }
+
+
+
+
+
 }
